Refactor

+++ controllers/answerController.js
@@ -3,7 +3,7 @@ const Answer = require('../models/answerModel')
 class AnswerController {
   static create (req, res) {
     let newAnswer = new Answer({
-      author: req.body.author, // req.decoded._id,
+      author: req.decoded._id,
       question: req.body.question,
       content: req.body.content
     })
@@ -16,9 +16,19 @@ class AnswerController {
     .catch(err => res.status(500).send(err))
   }
 
+  static findByQuestionId (req, res) {
+    Answer.find({ question: req.params.id })
+    .populate(['author'])
+    .then(answers => res.status(200).json({
+      message: 'Success find all answers',
+      answers: answers
+    }))
+    .catch(err => res.status(500).send(err))
+  }
+
   static findAll (req, res) {
     Answer.find()
-    .populate(['author', 'question'])
+    .populate(['author'])
     .then(answers => res.status(200).json({
       message: 'Success find all answers',
       answers: answers
@@ -28,11 +38,62 @@ class AnswerController {
 
   static findById (req, res) {
     Answer.findById(req.params.id)
+    .populate(['author'])
     .then(answer => res.status(200).json({
       message: 'Success find answer',
       answer: answer
     }))
     .catch(err => res.status(500).send(err))
+  }
+
+  static upVote(req, res) {
+    Answer.findById(req.params.id)
+    .then(answer => {
+      let userUpVoteIndex = answer.upVoters.findIndex(element => element == req.decoded._id)
+      let userDownVoteIndex = answer.downVoters.findIndex(element => element == req.decoded._id)
+
+      if (userUpVoteIndex === -1 && userDownVoteIndex === -1) { // false && false
+        answer.upVoters.push(req.decoded._id)
+      } else if (userUpVoteIndex === -1 && userDownVoteIndex !== -1) { // false && true
+        answer.downVoters.splice(userDownVoteIndex, 1)
+        answer.upVoters.push(req.decoded._id)
+      } else if (userUpVoteIndex !== -1 && userDownVoteIndex === -1) { // true && false
+        answer.upVoters.splice(userUpVoteIndex, 1)
+      }
+
+      answer.save()
+      .then(newAnswerData => res.status(200).json({
+        message: 'Success vote answer',
+        updatedAnswer: newAnswerData
+      }))
+      .catch(err => res.status(500).send(err))
+    })
+    .catch(err => res.status(500).send(err))
+  }
+
+  static downVote(req, res) {
+    Answer.findById(req.params.id)
+    .then(answer => {
+      let userUpVoteIndex = answer.upVoters.findIndex(element => element == req.decoded._id)
+      let userDownVoteIndex = answer.downVoters.findIndex(element => element == req.decoded._id)
+
+      if (userDownVoteIndex === -1 && userUpVoteIndex === -1) { // false && false
+        answer.downVoters.push(req.decoded._id)
+      } else if (userDownVoteIndex === -1 && userUpVoteIndex !== -1) { // false && true
+        answer.upVoters.splice(userUpVoteIndex, 1)
+        answer.downVoters.push(req.decoded._id)
+      } else if (userDownVoteIndex !== -1 && userUpVoteIndex === -1) { // true && false
+        answer.downVoters.splice(userDownVoteIndex, 1)
+      }
+
+      answer.save()
+        .then(newAnswerData => res.status(200).json({
+          message: 'Success vote answer',
+          updatedAnswer: newAnswerData
+        }))
+        .catch(err => res.status(500).send(err))
+      })
+      .catch(err => res.status(500).send(err))
   }
 
   static update (req, res) {
@@ -50,12 +111,20 @@ class AnswerController {
   }
 
   static delete (req, res) {
-    Answer.findByIdAndRemove(req.params.id)
-    .then(result => res.status(200).json({
-      message: 'Success delete answer',
-      deletedAnswer: result
-    }))
-    .catch(err => res.status(500).send(err))
+    Answer.findById(req.params.id)
+      .then(answer => {
+        if (answer.author == req.decoded._id) {
+          Answer.findByIdAndRemove(req.params.id)
+          .then(result => res.status(200).json({
+            message: 'Success delete answer',
+            deletedAnswer: result
+          }))
+          .catch(err => res.status(500).send(err))
+        } else {
+          res.status(403).send('Forbidden')
+        }
+      })
+      .catch(err => res.status(500).send(err))
   }
 }
